pyproject.toml
[project]
name = "rss-feed-aggregator"
version = "0.1.0"
description = "Whitelist-based RSS feed aggregator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
